PMTracker/requirements.txt
# Core Framework
pywebview==4.4.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
oracledb==2.0.0

# Text-to-Speech
gtts==2.4.0
pydub==0.25.1
pywin32==306

# Machine Learning
tensorflow==2.15.0
keras==2.15.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2

# Security & Utilities
cryptography==41.0.7
python-dotenv==1.0.0
pyperclip==1.8.2
openpyxl==3.1.2
reportlab==4.0.7
jinja2==3.1.2
requests==2.31.0
win10toast==0.9
